Add SOAP to REST API create page
